+++ game_patch/server/server.h
@@ -1,4 +1,7 @@
 #pragma once
 
+#include "../rf.h"
+
 void ServerInit();
 void ServerCleanup();
+bool CheckServerChatCommand(const char* msg, rf::Player* sender);