+++ CHANGELOG.md
@@ -3,6 +3,7 @@ DashFaction Changelog
 
 Version 1.3.1
 -------------
+- vote system (server-side)
 - rework high FPS jump fix - should fix free-fall animation bug and occasionally sticking to ground on FPS > 150
 - add 'map_ext' command for extending round time (server-side)
 - fix font texture leak (RF bug)